backend/issue-collector/requirements.txt
pydantic==2.11.9
pydantic-settings==2.11.0
SQLAlchemy==2.0.43
PyMySQL==1.1.1
requests==2.32.5
praw==7.8.1
python-dateutil==2.9.0.post0
APScheduler==3.11.0
typer==0.19.2
tzdata==2025.2
loguru==0.7.3
beautifulsoup4==4.13.5